+++ include/core/CompressUtils.h
