+++ lib/core/CompressUtils.cc
@@ -3,7 +3,7 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-#include <core/CCompressUtils.h>
+#include <core/CompressUtils.h>
 
 #include <core/CLogger.h>
