+++ lib/maths/CExpandingWindow.cc
@@ -6,7 +6,7 @@
 
 #include <maths/CExpandingWindow.h>
 
-#include <core/CCompressUtils.h>
+#include <core/CompressUtils.h>
 #include <core/CPersistUtils.h>
 #include <core/CStatePersistInserter.h>
 #include <core/CStateRestoreTraverser.h>
