+++ lib/model/CEventRateBucketGatherer.cc
@@ -6,7 +6,7 @@
 
 #include <model/CEventRateBucketGatherer.h>
 
-#include <core/CCompressUtils.h>
+#include <core/CompressUtils.h>
 #include <core/CFunctional.h>
 #include <core/CStatePersistInserter.h>
 #include <core/CStateRestoreTraverser.h>
